+++ src/Input/Base/AbstractHidden.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace PHPForge\Html\Input\Base;
 
 use PHPForge\Html\Attribute;
-use PHPForge\Html\Input\InputInterface;
+use PHPForge\Html\Input\Contract\InputInterface;
 use PHPForge\Html\Tag;
 use PHPForge\Widget\Element;
