+++ src/Input/Base/AbstractInputChoice.php
@@ -7,6 +7,7 @@ namespace PHPForge\Html\Input\Base;
 use PHPForge\Html\Attribute;
 use PHPForge\Html\Input\CheckedValueInterface;
 use PHPForge\Html\Input\ChoiceInterface;
+use PHPForge\Html\Input\Contract\ClassInterface;
 use PHPForge\Html\Label;
 use PHPForge\Html\Tag;
 use PHPForge\Widget\Element;
@@ -14,7 +15,7 @@ use PHPForge\Widget\Element;
 /**
  * Provides a foundation for creating HTML elements with various attributes and content.
  */
-abstract class AbstractInputChoice extends Element implements CheckedValueInterface, ChoiceInterface
+abstract class AbstractInputChoice extends Element implements CheckedValueInterface, ChoiceInterface, ClassInterface
 {
     use Attribute\Aria\HasAriaDescribedBy;
     use Attribute\Aria\HasAriaLabel;
